+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-04-25 20:57:11
-// Commit: test(i18n): implement widget integration tests for i18n
+// Updated at: 2026-04-25 22:22:07
+// Commit: chore(i18n): update dependencies and cleanup unused imports in i18n
 
 class i18nComponentState {
-  final String commitRef = '685';
-  final String lastMessage = 'test(i18n): implement widget integration tests for i18n';
-  final DateTime timestamp = DateTime.parse('2026-04-25T20:57:11+05:00');
+  final String commitRef = '693';
+  final String lastMessage = 'chore(i18n): update dependencies and cleanup unused imports in i18n';
+  final DateTime timestamp = DateTime.parse('2026-04-25T22:22:07+05:00');
 }